Rename/merge Kokkos_ArithmeticTraits.{hpp,cpp} -> Kokkos_NumericTraits.{hpp,cpp}
