+++ core/src/Kokkos_Core.hpp
@@ -55,7 +55,6 @@
 #include <Kokkos_Half.hpp>
 #include <KokkosCore_Config_DeclareBackend.hpp>
 
-#include <Kokkos_ArithmeticTraits.hpp>
 #include <Kokkos_AnonymousSpace.hpp>
 #include <Kokkos_LogicalSpaces.hpp>
 #include <Kokkos_Pair.hpp>
